// .temp/packages/types/generated.ts
export type AvatarUrls = {
    xs: string;
    sm: string;
    base: string;
    lg: string;
    xl: string;
    xxl: string;
};
export type User = {
    id: string;
    avatar_url: string;
    avatar_urls: AvatarUrls;
    cover_photo_url: string | null;
    email: string;
    username: string;
    display_name: string;
    system: boolean;
    integration: boolean;
    notifications_paused: boolean;
    notification_pause_expires_at: string | null;
    timezone: string | null;
    logged_in: boolean;
    type_name: string;
};
export type OrganizationMembershipStatus = {
    message: string;
    emoji: string;
    expiration_setting: '30m' | '1h' | '4h' | 'today' | 'this_week' | 'custom';
    expires_at: string | null;
    pause_notifications: boolean;
    expires_in: '30m' | '1h' | '4h' | 'today' | 'this_week' | 'custom';
};
export type OrganizationMember = {
    id: string;
    role: 'admin' | 'member' | 'viewer' | 'guest';
    created_at: string;
    deactivated: boolean;
    is_organization_member: boolean;
    user: User;
    status: OrganizationMembershipStatus | null;
};
// ...
export type SlackChannel = {
    id: string;
    name: string;
    is_private: boolean;
};
export type ProjectDisplayPreference = {
    display_reactions: boolean;
    display_attachments: boolean;
    display_comments: boolean;
    display_resolved: boolean;
};
export type Project = {
    id: string;
    name: string;
    description: string | null;
    created_at: string;
    archived_at: string | null;
    archived: boolean;
    last_activity_at: string;
    slack_channel_id: string | null;
    posts_count: number;
    cover_photo_url: string | null;
    url: string;
    accessory: string | null;
    private: boolean;
    personal: boolean;
    is_general: boolean;
    is_default: boolean;
    contributors_count: number;
    members_and_guests_count: number;
    members_count: number;
    guests_count: number;
    call_room_url: string | null;
    message_thread_id: string | null;
    organization_id: string;
    viewer_has_favorited: boolean;
    viewer_can_archive: boolean;
    viewer_can_destroy: boolean;
    viewer_can_unarchive: boolean;
    viewer_can_update: boolean;
    viewer_has_subscribed: boolean;
    viewer_subscription: 'posts_and_comments' | 'new_posts' | 'none';
    viewer_is_member: boolean;
    unread_for_viewer: boolean;
    slack_channel: SlackChannel | null;
    type_name: string;
    viewer_display_preferences: ProjectDisplayPreference | null;
    display_preferences: ProjectDisplayPreference;
};
// ...
export type ProjectMembership = {
    id: string;
    position: number;
    project: Project;
};
// ...
export type SyncUser = {
    id: string;
    avatar_urls: AvatarUrls;
    display_name: string;
    username: string;
    email: string;
    integration: boolean;
    notifications_paused: boolean;
};
export type SyncOrganizationMember = {
    id: string;
    role: 'admin' | 'member' | 'viewer' | 'guest';
    deactivated: boolean;
    last_seen_at: string | null;
    user: SyncUser;
};